# Conflicts:
#	Cargo.lock
#	crates/feagi-npu/runtime/src/embedded_impl/synapse_array.rs
#	crates/feagi-npu/runtime/src/std_impl/synapse_array.rs
#	crates/feagi-npu/runtime/src/traits/runtime.rs

+++ crates/feagi-npu/burst-engine/tests/gpu_shader_correctness_test.rs
@@ -58,8 +58,8 @@ fn create_test_genome(
             if synapse_idx < synapse_count {
                 synapse_array.source_neurons[synapse_idx] = source as u32;
                 synapse_array.target_neurons[synapse_idx] = target as u32;
-                synapse_array.weights[synapse_idx] = 128 + (i % 3) as u8 * 20;
-                synapse_array.postsynaptic_potentials[synapse_idx] = 200;
+                synapse_array.weights[synapse_idx] = 128.0 + (i % 3) as f32 * 20.0;
+                synapse_array.postsynaptic_potentials[synapse_idx] = 200.0;
                 synapse_array.types[synapse_idx] = if i % 4 == 0 { 1 } else { 0 }; // Mix exc/inh
                 synapse_array.valid_mask[synapse_idx] = true;
